+++ packages/credential-provider-ini/src/assumeRole/index.ts
@@ -1,0 +1,1 @@
+export * from "./resolveAssumeRoleCredentials";
