+++ packages/credential-provider-ini/src/credentialSource/index.ts
@@ -1,0 +1,1 @@
+export * from "./resolveCredentialSource";
