+++ packages/credential-provider-ini/src/profileData/index.ts
@@ -1,0 +1,1 @@
+export * from "./resolveProfileData";
