+++ packages/credential-provider-ini/src/sso/index.ts
@@ -1,0 +1,1 @@
+export * from "./resolveSsoCredentials";
