+++ packages/credential-provider-ini/src/staticCreds/index.ts
@@ -1,0 +1,1 @@
+export * from "./resolveStaticCredentials";
