+++ packages/credential-provider-ini/src/staticCreds/resolveStaticCredentials.spec.ts
