+++ packages/credential-provider-ini/src/staticCreds/resolveStaticCredentials.ts
