+++ packages/credential-provider-ini/src/webIdentity/index.ts
@@ -1,0 +1,1 @@
+export * from "./resolveWebIdentityCredentials";
