+++ packages/credential-provider-ini/src/webIdentity/resolveWebIdentityCredentials.spec.ts
